+++ Lorr/Editor/EditorApp.hh
@@ -12,11 +12,11 @@ struct EditorApp : Application {
     EditorLayout layout = {};
 
     bool prepare(this EditorApp &);
-    bool update(this EditorApp &, f32 delta_time);
+    bool update(this EditorApp &, f64 delta_time);
 
     bool do_super_init([[maybe_unused]] ls::span<c8 *> args) override { return true; };
     void do_shutdown() override {}
     bool do_prepare() override { return prepare(); }
-    bool do_update(f32 delta_time) override { return update(delta_time); }
+    bool do_update(f64 delta_time) override { return update(delta_time); }
 };
 }  // namespace lr
